==> notebooks/01_dataset_description_propertys.ipynb <==
import pandas as pd
from modlamp.descriptors import GlobalDescriptor

def get_mw(sequence):
    """Molecular Weight"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.calculate_MW(amide=True)
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_isoelectric_point(sequence):
    """Isoelectric point"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.isoelectric_point(amide=True)
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_charge_density(sequence):
    """Charge density"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.charge_density(ph=7, amide=True)
        return round(desc.descriptor[0][0], 5)
    except:
        return None

def get_charge(sequence):
    """Charge"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.calculate_charge(ph=7, amide=True)
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_instability_index(sequence):
    """Instability index"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.instability_index()
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_aromaticity(sequence):
    """Aromaticity"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.aromaticity()
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_aliphatic_index(sequence):
    """Aliphatic index"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.aliphatic_index()
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_boman_index(sequence):
    """Boman index"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.boman_index()
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_hydrophobic_ratio(sequence):
    """Hydrophobic ratio"""
    try:
        desc = GlobalDescriptor([sequence])
        desc.hydrophobic_ratio()
        return round(desc.descriptor[0][0], 4)
    except:
        return None

def get_frequency_aa(sequence):
    list_residues = ['A', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'N', 'K', 'L', 'M', 'P', 'Q', 'R', 'S', 'T', 'V', 'W', 'Y']
    row_description = [
        sequence.count(residue)/len(sequence)*100 for residue in list_residues
    ]

    return row_description

def coding_df(df_data, task):
    matrix_description =[]

    for index in df_data.index:
        sequence = df_data["sequence"][index]

        row = [
            sequence,
            get_mw(sequence),
            get_isoelectric_point(sequence),
            get_charge_density(sequence),
            get_charge(sequence),
            get_instability_index(sequence),
            get_aromaticity(sequence),
            get_aliphatic_index(sequence),
            get_boman_index(sequence),
            get_hydrophobic_ratio(sequence)
        ]

        row_count = get_frequency_aa(sequence)

        row = row + row_count
        matrix_description.append(row)

    df_description = pd.DataFrame(data=matrix_description, columns=["sequence", "Molecular weight", "Isoelectric Point",
                                                                    "Charge Density", "Charge", "Instability Index", "Aromaticity",
                                                                    "Aliphatic Index", "Boman Index", "Hydrophobic Ratio",
                                                                    'A', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'N', 'K', 'L', 'M', 
                                                                    'P', 'Q', 'R', 'S', 'T', 'V', 'W', 'Y'])

    df_description["Task"] = task
    return df_description

df_data = pd.read_csv("../results/dataset_fp.csv")
df_negative = df_data[df_data["monomer_state"] == 0]
df_positive = df_data[df_data["monomer_state"] == 1]

df_described_negative = coding_df(df_negative, "Negative data")
df_described_positive = coding_df(df_positive, "Positive data")

df_concat = pd.concat([df_described_negative, df_described_positive], axis=0)
df_concat.head()

df_concat.to_csv("../results/characterizing_dataset/described_dataset.csv", index=False)
