==> notebooks/01_dataset_preprocessing.ipynb <==
import pandas as pd

df = pd.read_csv("../input/osfp-full-data-set.csv")

df = df.rename(columns={"AminoAcidSequence": "sequence",
                        "OligomericState": "monomer_state"})

#### Change response to binary

df.loc[df["monomer_state"].str.contains("Dimer", case=False, na=False), "monomer_state"] = 0
df.loc[df["monomer_state"].str.contains("Tetramer", case=False, na=False), "monomer_state"] = 0

df.loc[df["monomer_state"].str.contains("Monomer", case=False, na=False), "monomer_state"] = 1

#### Verify that the residues are canonical

def is_aa_sequence(sequence):
    alphabet = set("ACDEFGHIKLMNPQRSTVWY")
    sequence = sequence.strip().upper()
    is_canon = True
    for res in set(sequence):
        if res not in alphabet:
            is_canon = False
    return is_canon

df["sequence"][0]

df["sequence"] = df["sequence"].str.replace(r"[\r\n]", "", regex=True)

df["is_canon"] = df["sequence"].apply(is_aa_sequence)

df["is_canon"].value_counts()

#### Balance of the response

df["monomer_state"].value_counts()

df = df[["sequence", "monomer_state"]]

#### Export dataset

df.to_csv("../results/dataset_fp.csv", index=False)
