==> notebooks/02_encoding_peptide_sequences.ipynb <==
### Numerical representation strategies demonstration

import pandas as pd
import sys
import os
sys.path.insert(0, "../")
from src.numerical_representation.physicochemical_properties import PhysicochemicalEncoder
from src.numerical_representation.embedding_representations import BioEmbeddings
from src.numerical_representation.fft_encoder import FFTTransform

### Loading dataset

df_data = pd.read_csv("../results/dataset_fp.csv")
df_data.head(5)

### Numerical representation strategies explored

#### Physicochemical properties

path_input = "../results/encoders"
dataset_encoder = pd.read_csv("../input_data_for_coding/aaindex_encoders.csv")

# ANDN920101 -> alpha helix
# ROBB760113 -> loop
# CRAJ730102 -> beta sheet
# ARGP820101 -> hydrophobicity

dataset_encoder = dataset_encoder[["residue", "ANDN920101", "ROBB760113", "ARGP820101", "CRAJ730102"]]
dataset_encoder.index = dataset_encoder["residue"]
dataset_encoder.head(5)

propertys = ["ANDN920101", "ROBB760113", "ARGP820101", "CRAJ730102"]

for group in propertys:
    command = f"mkdir -p {path_input}/physicochemical_properties/{group}"
    print(command)
    os.system(command)

    print("Start codifications")
    name_export = f"{path_input}/physicochemical_properties/{group}/coded_dataset.csv"

    physicochemical_encoder = PhysicochemicalEncoder(
        dataset=df_data,
        property_encoder = group,
        dataset_encoder=dataset_encoder,
        columns_to_ignore=["monomer_state"],
        name_column_seq="sequence"
    )

    physicochemical_encoder.run_process()
    physicochemical_encoder.df_data_encoded.to_csv(name_export, index=False)

#### FFT transform

for group in propertys:
    print("Reading datasets")
    df_coded = pd.read_csv(f"{path_input}/physicochemical_properties/{group}/coded_dataset.csv")

    command = f"mkdir -p {path_input}/FFT/{group}"
    print(command)
    os.system(command)

    print("Start codifications")

    name_export = f"{path_input}/FFT/{group}/coded_dataset.csv"

    print("Applying FFT")
    fft_transform = FFTTransform(
        dataset=df_coded,
        size_data=len(df_coded.columns)-1,
        columns_to_ignore=["monomer_state"],
    )

    response_coded = fft_transform.encoding_dataset()
    response_coded.to_csv(name_export, index=False)

#### Embedding through bio-embedding tool

bioembedding_instance = BioEmbeddings(
    dataset = df_data,
    seq_column = "sequence",
    is_reduced = True,
    device = "cuda",
    column_response = "monomer_state",
    path_export = "../results/encoders/"
)

bioembedding_instance.apply_bepler(name_export="coded_dataset")
bioembedding_instance.apply_onehot(name_export="coded_dataset") # One Hot
bioembedding_instance.apply_esm1b(name_export="coded_dataset")
bioembedding_instance.apply_esme(name_export="coded_dataset")
