==> notebooks/04_training_class_model.ipynb <==
## Training a classification model

import pandas as pd
import os
import sys
sys.path.insert(0, "../")

from src.training_models.classification_models import ClassificationModel
from sklearn.model_selection import train_test_split

### Load generated dataset

type_encoder = "FFT"
encoder = "ROBB760113"
df_data = pd.read_csv(f"../results/encoders/{type_encoder}/{encoder}/coded_dataset.csv")
df_data.head(5)

responses = df_data["monomer_state"].values
df_data = df_data.drop(columns=["monomer_state"])

train_data, validation_data, train_response, validation_response = train_test_split(df_data, responses, random_state=42, test_size=.30)

### Training model

folder_export = f"../results/performance/{type_encoder}/{encoder}"
os.makedirs(folder_export, exist_ok=True)
name_export = f"{folder_export}/results_exploration.csv"

model_instance = ClassificationModel(
    train_values = train_data,
    train_response = train_response,
    test_values = validation_data,
    test_response = validation_response,
    folder_export = folder_export
)

df_exploration = model_instance.apply_exploring()
df_exploration.to_csv(name_export, index=False)
